+++ src/useCases/services/DeleteServiceCategoryUseCase.ts
@@ -1,25 +1,36 @@
+import { ResourceNotFoundError } from "@/errors";
+import { checkUserPermissionForAsset } from "@/lib/auth/checkUserPermissionForAsset";
 import { IServiceCategoryRepository } from "@/repositories/IServiceCategoryRepository";
 
 interface DeleteServiceRequest {
   id: string;
   organizationId: string;
+  userId: string;
 }
 
 export class DeleteServiceCategoryUseCase {
   constructor(private serviceCategoryRepository: IServiceCategoryRepository) {}
 
-  async execute({ id, organizationId }: DeleteServiceRequest): Promise<void> {
-    // 1. Verificar propriedade antes de deletar
+  async execute({
+    id,
+    organizationId,
+    userId,
+  }: DeleteServiceRequest): Promise<void> {
     const existingService = await this.serviceCategoryRepository.findById(
       id,
-      organizationId
+      organizationId,
     );
 
     if (!existingService) {
-      throw new Error(
-        "Categoria de Serviço não encontrado ou você não tem permissão para removê-lo."
-      );
+      throw new ResourceNotFoundError("Categoria de Serviço não encontrado.",);
     }
+
+    await checkUserPermissionForAsset(
+      "serviceCategory",
+      userId,
+      existingService,
+      "DELETE",
+    );
 
     // 2. Deletar
     await this.serviceCategoryRepository.delete(id);
